+++ pyzoo/test/zoo/orca/data/conftest.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 import os
-from zoo import ZooContext
+from zoo.orca import OrcaContext
 
 import pytest
 
@@ -26,7 +26,7 @@ ray_ctx = None
 def orca_data_fixture():
     from zoo import init_spark_on_local
     from zoo.ray import RayContext
-    ZooContext._orca_eager_mode = True
+    OrcaContext.orca_eager_mode = True
     sc = init_spark_on_local(cores=4, spark_log_level="INFO")
     access_key_id = os.getenv("AWS_ACCESS_KEY_ID")
     secret_access_key = os.getenv("AWS_SECRET_ACCESS_KEY")
